+++ pkg/execute/factory.go
@@ -33,7 +33,7 @@ type DefaultExecutorFactory struct {
 	execExecutor          *ExecExecutor
 	sourceExecutor        *SourceExecutor
 	merger                *kubectl.Merger
-	cfgManager            ConfigPersistenceManager
+	cfgManager            config.PersistenceManager
 	kubectlCmdBuilder     *KubectlCmdBuilder
 	cmdsMapping           *CommandMapping
 	auditReporter         audit.AuditReporter
@@ -46,7 +46,7 @@ type DefaultExecutorFactoryParams struct {
 	Cfg               config.Config
 	KcChecker         *kubectl.Checker
 	Merger            *kubectl.Merger
-	CfgManager        ConfigPersistenceManager
+	CfgManager        config.PersistenceManager
 	AnalyticsReporter AnalyticsReporter
 	NamespaceLister   NamespaceLister
 	CommandGuard      CommandGuard
@@ -61,11 +61,11 @@ type Executor interface {
 }
 
 // ConfigPersistenceManager manages persistence of the configuration.
-type ConfigPersistenceManager interface {
-	PersistSourceBindings(ctx context.Context, commGroupName string, platform config.CommPlatformIntegration, channelAlias string, sourceBindings []string) error
-	PersistNotificationsEnabled(ctx context.Context, commGroupName string, platform config.CommPlatformIntegration, channelAlias string, enabled bool) error
-	PersistActionEnabled(ctx context.Context, name string, enabled bool) error
-}
+// type ConfigPersistenceManager interface {
+// 	PersistSourceBindings(ctx context.Context, commGroupName string, platform config.CommPlatformIntegration, channelAlias string, sourceBindings []string) error
+// 	PersistNotificationsEnabled(ctx context.Context, commGroupName string, platform config.CommPlatformIntegration, channelAlias string, enabled bool) error
+// 	PersistActionEnabled(ctx context.Context, name string, enabled bool) error
+// }
 
 // AnalyticsReporter defines a reporter that collects analytics data.
 type AnalyticsReporter interface {
@@ -179,7 +179,6 @@ func NewExecutorFactory(params DefaultExecutorFactoryParams) (*DefaultExecutorFa
 		execExecutor:          execExecutor,
 		sourceExecutor:        sourceExecutor,
 		merger:                params.Merger,
-		cfgManager:            params.CfgManager,
 		kubectlExecutor:       kcExecutor,
 		cmdsMapping:           mappings,
 		auditReporter:         params.AuditReporter,
@@ -226,7 +225,6 @@ func (f *DefaultExecutorFactory) NewDefault(cfg NewDefaultInput) Executor {
 		execExecutor:          f.execExecutor,
 		sourceExecutor:        f.sourceExecutor,
 		merger:                f.merger,
-		cfgManager:            f.cfgManager,
 		kubectlCmdBuilder:     f.kubectlCmdBuilder,
 		cmdsMapping:           f.cmdsMapping,
 		auditReporter:         f.auditReporter,
